Revert `Speed (1994)` to "dumb mode"
Calculating the max speed for certain vehicles isn't reliable - planes result in a lower max speed than some cars.

One idea would be to make an override map, but for now this should be enough.

The current speed calculations are also more accurate now, meaning it'll work for planes again.